#4 (ongoing) Add response client
In this commit a new mqtt client is added to handle only responses. This is done to have seperation of concerns since there are vast amount of mqtt messages to be handeled. Having one client responisble for acting on incoming messages(subscribing) and one for sending responses (publishing) divids responsibility in a simple way. I have theory about this being benificial during higher loads, we might be able to confirm or deny this if we load test later